webhook.py
from fastapi import FastAPI, Request, HTTPException
from pydantic import BaseModel
import logging

app = FastAPI()

# Create a Pydantic model to parse the incoming JSON data
class Prediction(BaseModel):
    id: str
    version: str
    created_at: str
    started_at: str = None
    completed_at: str = None
    status: str
    input: dict
    output: dict = None
    error: str = None
    logs: str = None
    metrics: dict = None

# Set up basic logging
logging.basicConfig(level=logging.INFO)

# Define the webhook endpoint
@app.post("/webhooks/replicate/face-to-many")
async def replicate_webhook(prediction: Prediction):
    logging.info(f"🪝 Incoming webhook! Prediction ID: {prediction.id}")

    # # Process the prediction object
    # await save_to_my_database(prediction)
    # await send_slack_notification(prediction)

    # Respond with a 200 status code to acknowledge receipt of the webhook
    return {"message": "Webhook received successfully"}

# # Mock function to save the prediction data to a database
# async def save_to_my_database(prediction: Prediction):
#     logging.info(f"Saving prediction ID: {prediction.id} to the database")
#     # Your database saving logic goes here
#     pass

# # Mock function to send a Slack notification
# async def send_slack_notification(prediction: Prediction):
#     logging.info(f"Sending Slack notification for prediction ID: {prediction.id}")
#     # Your Slack notification logic goes here
#     pass

