requirements.txt
TgCrypto
pyrogram
pillow
requests
psycopg2-binary
schedule
replicate
fastapi
